requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
httpx==0.28.1
python-dotenv==1.1.1
msal==1.32.3
reportlab==4.4.3
pydantic==2.11.7
